move commands to their locations (#110714)
