+++ src/vs/workbench/api/common/apiCommands.ts
@@ -9,11 +9,9 @@ import * as typeConverters from 'vs/workbench/api/common/extHostTypeConverters';
 import { CommandsRegistry, ICommandService, ICommandHandler } from 'vs/platform/commands/common/commands';
 import { ITextEditorOptions } from 'vs/platform/editor/common/editor';
 import { EditorGroupColumn } from 'vs/workbench/common/editor';
-import { EditorGroupLayout } from 'vs/workbench/services/editor/common/editorGroupsService';
 import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
-import { IOpenWindowOptions, IWindowOpenable, IOpenEmptyWindowOptions } from 'vs/platform/windows/common/windows';
-import { IWorkspacesService, hasWorkspaceFileExtension, IRecent } from 'vs/platform/workspaces/common/workspaces';
-import { Schemas } from 'vs/base/common/network';
+import { IOpenEmptyWindowOptions } from 'vs/platform/windows/common/windows';
+import { IWorkspacesService, IRecent } from 'vs/platform/workspaces/common/workspaces';
 import { ILogService } from 'vs/platform/log/common/log';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
 import { IViewDescriptorService, IViewsService, ViewVisibilityState } from 'vs/workbench/common/views';
@@ -34,41 +32,6 @@ function adjustHandler(handler: (executor: ICommandsExecutor, ...args: any[]) =>
 		return handler(accessor.get(ICommandService), ...args);
 	};
 }
-
-interface IOpenFolderAPICommandOptions {
-	forceNewWindow?: boolean;
-	forceReuseWindow?: boolean;
-	noRecentEntry?: boolean;
-}
-
-export class OpenFolderAPICommand {
-	public static readonly ID = 'vscode.openFolder';
-	public static execute(executor: ICommandsExecutor, uri?: URI, forceNewWindow?: boolean): Promise<any>;
-	public static execute(executor: ICommandsExecutor, uri?: URI, options?: IOpenFolderAPICommandOptions): Promise<any>;
-	public static execute(executor: ICommandsExecutor, uri?: URI, arg: boolean | IOpenFolderAPICommandOptions = {}): Promise<any> {
-		if (typeof arg === 'boolean') {
-			arg = { forceNewWindow: arg };
-		}
-		if (!uri) {
-			return executor.executeCommand('_files.pickFolderAndOpen', { forceNewWindow: arg.forceNewWindow });
-		}
-		const options: IOpenWindowOptions = { forceNewWindow: arg.forceNewWindow, forceReuseWindow: arg.forceReuseWindow, noRecentEntry: arg.noRecentEntry };
-		uri = URI.revive(uri);
-		const uriToOpen: IWindowOpenable = (hasWorkspaceFileExtension(uri) || uri.scheme === Schemas.untitled) ? { workspaceUri: uri } : { folderUri: uri };
-		return executor.executeCommand('_files.windowOpen', [uriToOpen], options);
-	}
-}
-CommandsRegistry.registerCommand({
-	id: OpenFolderAPICommand.ID,
-	handler: adjustHandler(OpenFolderAPICommand.execute),
-	description: {
-		description: 'Open a folder or workspace in the current window or new window depending on the newWindow argument. Note that opening in the same window will shutdown the current extension host process and start a new one on the given folder/workspace unless the newWindow parameter is set to true.',
-		args: [
-			{ name: 'uri', description: '(optional) Uri of the folder or workspace file to open. If not provided, a native dialog will ask the user for the folder', constraint: (value: any) => value === undefined || value instanceof URI },
-			{ name: 'options', description: '(optional) Options. Object with the following properties: `forceNewWindow `: Whether to open the folder/workspace in a new window or the same. Defaults to opening in the same window. `noRecentEntry`: Wheter the opened URI will appear in the \'Open Recent\' list. Defaults to true.  Note, for backward compatibility, options can also be of type boolean, representing the `forceNewWindow` setting.', constraint: (value: any) => value === undefined || typeof value === 'object' || typeof value === 'boolean' }
-		]
-	}
-});
 
 interface INewWindowAPICommandOptions {
 	reuseWindow?: boolean;
@@ -178,38 +141,6 @@ CommandsRegistry.registerCommand('_workbench.addToRecentlyOpened', async functio
 CommandsRegistry.registerCommand('_workbench.getRecentlyOpened', async function (accessor: ServicesAccessor) {
 	const workspacesService = accessor.get(IWorkspacesService);
 	return workspacesService.getRecentlyOpened();
-});
-
-export class SetEditorLayoutAPICommand {
-	public static readonly ID = 'vscode.setEditorLayout';
-	public static execute(executor: ICommandsExecutor, layout: EditorGroupLayout): Promise<any> {
-		return executor.executeCommand('layoutEditorGroups', layout);
-	}
-}
-CommandsRegistry.registerCommand({
-	id: SetEditorLayoutAPICommand.ID,
-	handler: adjustHandler(SetEditorLayoutAPICommand.execute),
-	description: {
-		description: 'Set Editor Layout',
-		args: [{
-			name: 'args',
-			schema: {
-				'type': 'object',
-				'required': ['groups'],
-				'properties': {
-					'orientation': {
-						'type': 'number',
-						'default': 0,
-						'enum': [0, 1]
-					},
-					'groups': {
-						'$ref': '#/definitions/editorGroupsSchema', // defined in keybindingService.ts ...
-						'default': [{}, {}],
-					}
-				}
-			}
-		}]
-	}
 });
 
 CommandsRegistry.registerCommand('_extensionTests.setLogLevel', function (accessor: ServicesAccessor, level: number) {
